# Net: Tiny speedup, less allocation