--- a/data/transactions/verify/verifiedTxnCache.go
+++ b/data/transactions/verify/verifiedTxnCache.go
@@ -49,7 +49,7 @@ var errTooManyPinnedEntries = &VerifiedTxnCacheError{errors.New("Too many pinned
 // errMissingPinnedEntry is being generated when we're trying to pin a transaction that does not appear in the cache
 var errMissingPinnedEntry = &VerifiedTxnCacheError{errors.New("Missing pinned entry")}
 
-// VerifiedTransactionCache provides a cached store of recently verified transactions. The cache is desiged to have two separate "levels". On the
+// VerifiedTransactionCache provides a cached store of recently verified transactions. The cache is designed to have two separate "levels". On the
 // bottom tier, the cache would be using a cyclic buffer, where old transactions would end up overridden by new ones. In order to support transactions
 // that goes into the transaction pool, we have a higher tier of pinned cache. Pinned transactions would not be cycled-away by new incoming transactions,
 // and would only get eliminated by updates to the transaction pool, which would inform the cache of updates to the pinned items.