Very small twekas while reading txHandler.go

This makes the saltedCache a tiny bit faster, and perform fewer allocs.
```
                                                  │ before.txt  │             after.txt              │
                                                  │   sec/op    │   sec/op     vs base               │
DigestCaches/data.digestCacheMaker/threads=1-10     720.7n ± 1%   723.9n ± 1%       ~ (p=0.210 n=10)
DigestCaches/data.saltedCacheMaker/threads=1-10     988.0n ± 1%   979.1n ± 1%  -0.90% (p=0.015 n=10)
DigestCaches/data.digestCacheMaker/threads=4-10     764.9n ± 1%   763.1n ± 1%       ~ (p=0.403 n=10)
DigestCaches/data.saltedCacheMaker/threads=4-10     1.319µ ± 0%   1.309µ ± 0%  -0.80% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=16-10    997.5n ± 1%   994.1n ± 1%       ~ (p=0.305 n=10)
DigestCaches/data.saltedCacheMaker/threads=16-10    1.325µ ± 0%   1.309µ ± 1%  -1.21% (p=0.002 n=10)
DigestCaches/data.digestCacheMaker/threads=128-10   1.023µ ± 1%   1.020µ ± 0%       ~ (p=0.147 n=10)
DigestCaches/data.saltedCacheMaker/threads=128-10   1.356µ ± 0%   1.336µ ± 0%  -1.47% (p=0.000 n=10)
geomean                                             1.035µ        1.029µ       -0.61%

                                                  │ before.txt  │              after.txt               │
                                                  │    B/op     │    B/op     vs base                  │
DigestCaches/data.digestCacheMaker/threads=1-10     173.0 ±  2%   173.5 ± 0%        ~ (p=0.331 n=10)
DigestCaches/data.saltedCacheMaker/threads=1-10     171.0 ±  1%   138.0 ± 1%  -19.30% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=4-10     184.0 ±  1%   184.0 ± 1%        ~ (p=0.914 n=10)
DigestCaches/data.saltedCacheMaker/threads=4-10     189.5 ±  1%   156.0 ± 1%  -17.68% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=16-10    140.0 ± 11%   139.0 ± 1%        ~ (p=0.138 n=10)
DigestCaches/data.saltedCacheMaker/threads=16-10    189.5 ±  1%   156.0 ± 1%  -17.68% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=128-10   156.0 ±  0%   156.0 ± 0%        ~ (p=1.000 n=10) ¹
DigestCaches/data.saltedCacheMaker/threads=128-10   193.5 ±  1%   159.0 ± 1%  -17.83% (p=0.000 n=10)
geomean                                             173.6         157.0        -9.56%
¹ all samples are equal

                                                  │ before.txt │              after.txt               │
                                                  │ allocs/op  │ allocs/op   vs base                  │
DigestCaches/data.digestCacheMaker/threads=1-10     1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
DigestCaches/data.saltedCacheMaker/threads=1-10     2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=4-10     1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
DigestCaches/data.saltedCacheMaker/threads=4-10     2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=16-10    1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
DigestCaches/data.saltedCacheMaker/threads=16-10    2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
DigestCaches/data.digestCacheMaker/threads=128-10   1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
DigestCaches/data.saltedCacheMaker/threads=128-10   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
geomean                                             1.414        1.000       -29.29%
```
<!--
Thanks for submitting a pull request! We appreciate the time and effort you spent to get this far.

If you haven't already, please make sure that you've reviewed the CONTRIBUTING guide:
https://github.com/algorand/go-algorand/blob/master/CONTRIBUTING.md#code-guidelines

In particular ensure that you've run the following:
* make generate
* make sanity (which runs make fmt, make lint, make fix and make vet)

It is also a good idea to run tests:
* make test
* make integration
-->

## Summary

<!-- Explain the goal of this change and what problem it is solving. Format this cleanly so that it may be used for a commit message, as your changes will be squash-merged. -->

## Test Plan

<!-- How did you test these changes? Please provide the exact scenarios you tested in as much detail as possible including commands, output and rationale. -->